fix(auth): add lib/middleware/protect and refactor some codes
