+++ src/app/api/v1/group/score/[id]/route.ts
@@ -1,6 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
-import { db, firebaseAdmin } from "@/lib/services/firebase.admin";
+import { db } from "@/lib/services/firebase.admin";
 import { FieldValue } from "firebase-admin/firestore";
+import { protect } from "@/lib/middleware/protect";
 
 // Get each group score : Public
 export async function GET(
@@ -28,19 +29,9 @@ export async function POST(
   request: NextRequest,
   { params }: { params: Promise<{ id: string }> }
 ) {
-  // get token from the authorization request header
-  const authHeader = request.headers.get("Authorization");
-  if (!authHeader || !authHeader.startsWith("Bearer ")) {
-    return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
-  }
-
-  const token = authHeader.split(" ")[1];
-
-  //decoding the token to get the user email and role
-  const session = await firebaseAdmin.auth().verifyIdToken(token);
-  const role: string = session?.role;
-  if (role != "admin") {
-    return NextResponse.json({ error: "Invalid credentials" }, { status: 403 });
+  const res = await protect(request, ["admin"]);
+  if (res) {
+    return res;
   }
 
   // get body
@@ -70,21 +61,10 @@ export async function PUT(
   request: NextRequest,
   { params }: { params: Promise<{ id: string }> }
 ) {
-  // get token from the authorization request header
-  const authHeader = request.headers.get("Authorization");
-  if (!authHeader || !authHeader.startsWith("Bearer ")) {
-    return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
+  const res = await protect(request, ["admin"]);
+  if (res) {
+    return res;
   }
-
-  const token = authHeader.split(" ")[1];
-
-  //decoding the token to get the user email and role
-  const session = await firebaseAdmin.auth().verifyIdToken(token);
-  const role: string = session?.role;
-  if (role != "admin") {
-    return NextResponse.json({ error: "Invalid credentials" }, { status: 403 });
-  }
-
   // get body
   const body = await request.json();
   const { newScore } = body;
